Fix bug 1273757: Add rate limit for newsletter subscriptions
* Rate limit is per newsletter(s) per email address.
* Add and fix tests for rate limiting email subscription
* Fix update_user tests
* Move tests for update_user_task to a separate file
* Rename test file to upsert_user to match function
